Add files via upload
The Trivial File Transfer Protocol (TFTP) is a simple way of transferring file between
two systems. This protocol is specified in RFC 1350.This protocol doesn’t support
advanced available in FTP (ex: User authentication), typically used in Embedded
systems for its smaller footprint and simplicity. The goal of this project is to understand
RFC and implement the protocol in a LAN environment. Eventually this project can be
extended for advanced cases like inter-operating with standard TFTP client / servers.